tools / adding portable apps
as some people claimed they cannot or would like not to install applications
or enable preparing a USB stick ready for editing/reading/working with the books, e.g. when writing sth. yourself on sources

also it is easy to integrate e.g. TeXlive and TeXstudio for your LaTeX works with that as shown on the [Uni-Stick](https://wiki.fsfw-dresden.de/doku.php/doku/uni-stick).